src/webpubsub/python/requirements.txt
Flask==3.0.0
llama_cpp_python==0.2.18
azure-messaging-webpubsubservice==1.1.0b1
azure-messaging-webpubsubclient==1.0.0b1